Move LazyGroupBy to submodule
